Create reusable NewCollectionDialog component and integrate with both Welcome Page and Configurations List
- Created NewCollectionDialog.vue component with full collection creation logic
- Component includes collection name, description, and version components
- Proper test data file naming with enumerators version (.json extension)
- Integrated component into Welcome Page (replacing inline dialog)
- Integrated component into Configurations List page (replacing inline dialog)
- Both pages now use identical collection creation workflow
- Removed duplicate code and logic from both pages
- Component emits 'created' event with file name for navigation
- Maintains all existing functionality while improving code organization
- Ensures consistent behavior and styling across both locations
