+++ src/pages/WelcomePage.vue
@@ -86,182 +86,11 @@
     />
 
     <!-- New Collection Dialog -->
-    <v-dialog v-model="showNewCollectionDialog" max-width="600px" data-test="new-collection-dialog">
-      <v-card>
-        <v-card-title class="text-h5 pa-6 pb-4" data-test="new-collection-dialog-title">
-          <v-icon start color="primary" class="mr-2">mdi-plus-circle</v-icon>
-          Create New Collection
-        </v-card-title>
-        <v-card-text class="pa-6 pt-0">
-          <div class="d-flex flex-column gap-6">
-            <!-- Collection Name -->
-            <div>
-              <h4 class="text-subtitle-1 font-weight-medium mb-4 text-medium-emphasis">Collection Details</h4>
-              <v-text-field
-                v-model="newCollectionName"
-                label="Collection Name"
-                :error="!!nameError"
-                :error-messages="nameError || undefined"
-                placeholder="my_collection"
-                :disabled="creating"
-                @keyup.enter="createCollection"
-                variant="outlined"
-                density="compact"
-                data-test="new-collection-name-input"
-              />
-              <p class="text-caption text-medium-emphasis mt-2" data-test="new-collection-help-text">
-                Collection names must start with a letter and contain only letters, numbers, and underscores.
-              </p>
-            </div>
-
-            <!-- Collection Description -->
-            <div>
-              <v-text-field
-                v-model="newCollectionDescription"
-                label="Collection Description"
-                placeholder="Describe the purpose of this collection"
-                :disabled="creating"
-                variant="outlined"
-                density="compact"
-                data-test="new-collection-description-input"
-              />
-            </div>
-
-            <!-- Version Components -->
-            <div class="version-components">
-              <h4 class="text-subtitle-1 font-weight-medium mb-4 text-medium-emphasis">Version Components</h4>
-              
-              <!-- Major Version -->
-              <div class="version-row mb-4">
-                <div class="version-label">
-                  <span class="text-body-1 font-weight-medium">Major</span>
-                  <span class="text-caption text-medium-emphasis d-block">Breaking changes</span>
-                </div>
-                <div class="version-controls">
-                  <v-text-field
-                    v-model.number="newVersion.major"
-                    type="number"
-                    variant="outlined"
-                    density="compact"
-                    hide-details
-                    class="version-input"
-                    min="0"
-                  />
-                  <v-btn
-                    icon="mdi-plus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.major++"
-                    :disabled="creating"
-                  />
-                  <v-btn
-                    icon="mdi-minus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.major = Math.max(0, newVersion.major - 1)"
-                    :disabled="creating"
-                  />
-                </div>
-              </div>
-
-              <!-- Minor Version -->
-              <div class="version-row mb-4">
-                <div class="version-label">
-                  <span class="text-body-1 font-weight-medium">Minor</span>
-                  <span class="text-caption text-medium-emphasis d-block">New features</span>
-                </div>
-                <div class="version-controls">
-                  <v-text-field
-                    v-model.number="newVersion.minor"
-                    type="number"
-                    variant="outlined"
-                    density="compact"
-                    hide-details
-                    class="version-input"
-                    min="0"
-                  />
-                  <v-btn
-                    icon="mdi-plus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.minor++"
-                    :disabled="creating"
-                  />
-                  <v-btn
-                    icon="mdi-minus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.minor = Math.max(0, newVersion.minor - 1)"
-                    :disabled="creating"
-                  />
-                </div>
-              </div>
-
-              <!-- Patch Version -->
-              <div class="version-row mb-4">
-                <div class="version-label">
-                  <span class="text-body-1 font-weight-medium">Patch</span>
-                  <span class="text-caption text-medium-emphasis d-block">Bug fixes</span>
-                </div>
-                <div class="version-controls">
-                  <v-text-field
-                    v-model.number="newVersion.patch"
-                    type="number"
-                    variant="outlined"
-                    density="compact"
-                    hide-details
-                    class="version-input"
-                    min="0"
-                  />
-                  <v-btn
-                    icon="mdi-plus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.patch++"
-                    :disabled="creating"
-                  />
-                  <v-btn
-                    icon="mdi-minus"
-                    variant="text"
-                    size="small"
-                    @click="newVersion.patch = Math.max(0, newVersion.patch - 1)"
-                    :disabled="creating"
-                  />
-                </div>
-              </div>
-
-              <!-- Version Preview -->
-              <div class="version-preview">
-                <div class="text-caption text-medium-emphasis mb-2">Final Version:</div>
-                <div class="version-display">
-                  <span class="text-h6 font-weight-bold">{{ newVersionString }}</span>
-                  <span class="text-caption text-medium-emphasis d-block">+ enumerator version</span>
-                </div>
-              </div>
-            </div>
-          </div>
-        </v-card-text>
-        <v-card-actions class="pa-6 pt-0">
-          <v-spacer />
-          <v-btn
-            @click="showNewCollectionDialog = false"
-            :disabled="creating"
-            data-test="new-collection-cancel-btn"
-          >
-            Cancel
-          </v-btn>
-          <v-btn
-            color="primary"
-            @click="createCollection"
-            :loading="creating"
-            :disabled="!newCollectionName.trim()"
-            data-test="new-collection-create-btn"
-          >
-            Create
-          </v-btn>
-        </v-card-actions>
-      </v-card>
-    </v-dialog>
+    <NewCollectionDialog
+      v-model="showNewCollectionDialog"
+      @created="handleCollectionCreated"
+      data-test="new-collection-dialog"
+    />
   </v-container>
 </template>
 
@@ -271,6 +100,7 @@ import { useRoute, useRouter } from 'vue-router'
 import { useHelp } from '@/composables/useHelp'
 import { apiService } from '@/utils/api'
 import HelpDialog from '@/components/HelpDialog.vue'
+import NewCollectionDialog from '@/components/NewCollectionDialog.vue'
 
 const { showHelp, currentHelp } = useHelp()
 const router = useRouter()
@@ -280,21 +110,6 @@ const route = useRoute()
 
 // New collection dialog state
 const showNewCollectionDialog = ref(false)
-const newCollectionName = ref('')
-const newCollectionDescription = ref('')
-const nameError = ref<string | null>(null)
-const creating = ref(false)
-
-// Version state for new collection
-const newVersion = ref({
-  major: 0,
-  minor: 0,
-  patch: 0,
-})
-
-const newVersionString = computed(() => {
-  return `${newVersion.value.major}.${newVersion.value.minor}.${newVersion.value.patch}`
-})
 
 // Initialize slide based on URL parameter
 onMounted(() => {
@@ -307,111 +122,14 @@ onMounted(() => {
   }
 })
 
-// Validate collection name
-const validateCollectionName = (name: string): boolean => {
-  const pattern = /^[a-zA-Z][a-zA-Z0-9_]*$/
-  if (!pattern.test(name)) {
-    nameError.value = 'Collection name must start with a letter and contain only letters, numbers, and underscores'
-    return false
-  }
-  nameError.value = null
-  return true
-}
-
-// Create new collection using PUT endpoints
-const createCollection = async () => {
-  const name = newCollectionName.value.trim()
-  
-  if (!name) {
-    nameError.value = 'Collection name is required'
-    return
-  }
-
-  if (!validateCollectionName(name)) {
-    return
-  }
-
-  creating.value = true
-  try {
-    // Get the newest enumerator version
-    const enumeratorFiles = await apiService.getEnumerators()
-    let newestEnumeratorVersion = 0
-    
-    if (enumeratorFiles && enumeratorFiles.length > 0) {
-      // Filter out any files that don't have the expected structure
-      const validFiles = enumeratorFiles.filter((f: any) => f && f.file_name && typeof f.file_name === 'string')
-      
-      if (validFiles.length > 0) {
-        const versions = validFiles
-          .map((f: any) => {
-            const match = f.file_name.match(/enumerations\.(\d+)\.yaml/)
-            return match ? parseInt(match[1], 10) : 0
-          })
-          .filter((version: number) => !isNaN(version))
-        
-        newestEnumeratorVersion = versions.length > 0 ? Math.max(...versions) : 0
-      }
-    }
-    
-    const version = `${newVersion.value.major}.${newVersion.value.minor}.${newVersion.value.patch}.${newestEnumeratorVersion}`
-    const configFileName = `${name}.yaml`
-    
-    // Create configuration file
-    const configuration = {
-      file_name: configFileName,
-      title: name,
-      description: newCollectionDescription.value || `Configuration for ${name} collection`,
-      versions: [{
-        version: version,
-        dictionary: `${name}.${newVersion.value.major}.${newVersion.value.minor}.${newVersion.value.patch}.yaml`,
-        enumerators: `enumerations.${newestEnumeratorVersion}.yaml`,
-        indexes: [],
-        migrations: []
-      }]
-    }
-    
-    // Create dictionary file
-    const dictionary = {
-      root: {
-        name: "root",
-        type: "object",
-        description: newCollectionDescription.value || `Dictionary for ${name} collection`,
-        properties: [
-          {"name": "_id", "type": "identifier"},
-          {"name": "name", "type": "word"},
-          {"name": "last_saved", "type": "breadcrumb"}
-        ]
-      }
-    }
-    
-    // Create test data file
-    const testData: any[] = []
-    
-    // Save all files using PUT endpoints
-    await apiService.saveConfiguration(configFileName, configuration)
-    await apiService.saveDictionary(`${name}.${newVersion.value.major}.${newVersion.value.minor}.${newVersion.value.patch}.yaml`, dictionary)
-    await apiService.saveTestDataFile(`${name}.${newVersion.value.major}.${newVersion.value.minor}.${newVersion.value.patch}.${newestEnumeratorVersion}.json`, testData)
-    
-    // Close dialog and reset
-    showNewCollectionDialog.value = false
-    newCollectionName.value = ''
-    newCollectionDescription.value = ''
-    nameError.value = null
-    newVersion.value = { major: 0, minor: 0, patch: 0 }
-    
-    // Navigate to the new collection's detail page
-    router.push(`/configurations/${configFileName}`)
-    
-  } catch (err: any) {
-    nameError.value = err.message || 'Failed to create collection'
-  } finally {
-    creating.value = false
-  }
-}
-
 // Open new collection dialog
 const createNewCollection = () => {
   showNewCollectionDialog.value = true
+}
+
+// Handle collection creation from NewCollectionDialog
+const handleCollectionCreated = (fileName: string) => {
+  router.push(`/configurations/${fileName}`)
 }
 
 const helpSlides = [
@@ -655,42 +373,20 @@ const nextSlide = () => {
 <style scoped>
 .help-carousel {
   height: 100vh;
-  border: none;
-  border-radius: 0;
-}
-
-.help-window {
-  height: calc(100vh - 120px);
-}
-
-.help-window :deep(.v-window__container) {
-  height: 100%;
-}
-
-.help-window :deep(.v-window-item) {
-  height: 100%;
+  overflow: hidden;
 }
 
 .carousel-container {
   position: relative;
-  display: flex;
-  align-items: center;
-  height: calc(100vh - 120px);
-  padding: 0 80px; /* Add padding to account for navigation buttons */
+  height: 100%;
 }
 
 .carousel-nav-btn {
   position: absolute;
+  top: 50%;
+  transform: translateY(-50%);
   z-index: 10;
-  background-color: rgba(255, 255, 255, 0.9) !important;
-  backdrop-filter: blur(10px);
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-}
-
-.carousel-nav-btn:hover {
-  background-color: rgba(255, 255, 255, 1) !important;
-  transform: scale(1.1);
+  background-color: rgba(255, 255, 255, 0.9);
 }
 
 .prev-btn {
@@ -701,95 +397,15 @@ const nextSlide = () => {
   right: 20px;
 }
 
-.carousel-nav-btn:disabled {
-  opacity: 0.3;
-  cursor: not-allowed;
+.help-window {
+  height: 100%;
 }
 
 .overview-content {
-  width: 100%;
-  text-align: left;
-  overflow-y: auto;
-  max-height: calc(100vh - 300px);
-  padding: 0 20px;
+  max-width: 800px;
 }
 
-.detailed-content {
-  text-align: left;
-  line-height: 1.6;
-  width: 100%;
-  font-size: 1.1rem;
-}
-
-.detailed-content :deep(h2) {
-  font-size: 1.8rem;
-  font-weight: 600;
-  margin: 1.5rem 0 0.75rem 0;
+.detailed-content h2 {
   color: #1976d2;
-}
-
-.detailed-content :deep(h3) {
-  font-size: 1.5rem;
-  font-weight: 500;
-  margin: 1.25rem 0 0.5rem 0;
-  color: #424242;
-}
-
-.detailed-content :deep(p) {
-  margin-bottom: 0.75rem;
-  line-height: 1.6;
-  font-size: 1.1rem;
-}
-
-.detailed-content :deep(ul), .detailed-content :deep(ol) {
-  margin-bottom: 0.75rem;
-  padding-left: 1.5rem;
-  font-size: 1.1rem;
-}
-
-.detailed-content :deep(li) {
-  margin-bottom: 0.25rem;
-  font-size: 1.1rem;
-}
-
-.detailed-content :deep(strong) {
-  font-weight: 600;
-  color: #1976d2;
-}
-
-.version-components {
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  padding: 16px;
-  background-color: #fafafa;
-}
-
-.version-row {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-}
-
-.version-label {
-  flex: 1;
-}
-
-.version-controls {
-  display: flex;
-  align-items: center;
-}
-
-.version-input {
-  width: 80px;
-}
-
-.version-preview {
-  text-align: center;
-}
-
-.version-display {
-  padding: 16px;
-  background-color: #f5f5f5;
-  border-radius: 8px;
 }
 </style>
